requirements.txt
numpy==1.26.4
pyarrow==13.0.0  # Downgrade further for NumPy 1.x compatibility
streamlit==1.32.2
pandas==2.0.3
# openai==1.30.1
ollama
loguru==0.7.2
python-dotenv==1.0.1
protobuf==3.20.3 